+++ src/analyze-action.ts
@@ -19,22 +19,19 @@ import { runAutobuild } from "./autobuild";
 import { getCodeQL } from "./codeql";
 import { Config, getConfig } from "./config-utils";
 import { uploadDatabases } from "./database-upload";
+import { EnvVar } from "./environment";
 import { Features } from "./feature-flags";
 import { Language } from "./languages";
 import { getActionsLogger, Logger } from "./logging";
 import { parseRepositoryNwo } from "./repository";
-import {
-  CODEQL_ACTION_ANALYZE_DID_COMPLETE_SUCCESSFULLY,
-  CODEQL_ACTION_DID_AUTOBUILD_GOLANG,
-} from "./shared-environment";
 import { getTotalCacheSize, uploadTrapCaches } from "./trap-caching";
-import * as upload_lib from "./upload-lib";
+import * as uploadLib from "./upload-lib";
 import { UploadResult } from "./upload-lib";
 import * as util from "./util";
 import { checkForTimeout, wrapError } from "./util";
 
 interface AnalysisStatusReport
-  extends upload_lib.UploadStatusReport,
+  extends uploadLib.UploadStatusReport,
     QueriesStatusReport {}
 
 interface FinishStatusReport
@@ -144,7 +141,7 @@ async function runAutobuildIfLegacyGoWorkflow(config: Config, logger: Logger) {
   if (!config.languages.includes(Language.go)) {
     return;
   }
-  if (process.env[CODEQL_ACTION_DID_AUTOBUILD_GOLANG] === "true") {
+  if (process.env[EnvVar.DID_AUTOBUILD_GOLANG] === "true") {
     logger.debug("Won't run Go autobuild since it has already been run.");
     return;
   }
@@ -213,9 +210,6 @@ async function run() {
       actionsUtil.getOptionalInput("threads") || process.env["CODEQL_THREADS"],
       logger
     );
-    const memory = util.getMemoryFlag(
-      actionsUtil.getOptionalInput("ram") || process.env["CODEQL_RAM"]
-    );
 
     const repositoryNwo = parseRepositoryNwo(
       util.getRequiredEnvParam("GITHUB_REPOSITORY")
@@ -228,6 +222,11 @@ async function run() {
       repositoryNwo,
       actionsUtil.getTemporaryDirectory(),
       logger
+    );
+
+    const memory = await util.getMemoryFlag(
+      actionsUtil.getOptionalInput("ram") || process.env["CODEQL_RAM"],
+      features
     );
 
     await runAutobuildIfLegacyGoWorkflow(config, logger);
@@ -269,7 +268,7 @@ async function run() {
     core.setOutput("db-locations", dbLocations);
     const uploadInput = actionsUtil.getOptionalInput("upload");
     if (runStats && actionsUtil.getUploadValue(uploadInput) === "always") {
-      uploadResult = await upload_lib.uploadFromActions(
+      uploadResult = await uploadLib.uploadFromActions(
         outputDir,
         actionsUtil.getRequiredInput("checkout_path"),
         actionsUtil.getOptionalInput("category"),
@@ -296,7 +295,7 @@ async function run() {
       uploadResult !== undefined &&
       actionsUtil.getRequiredInput("wait-for-processing") === "true"
     ) {
-      await upload_lib.waitForProcessing(
+      await uploadLib.waitForProcessing(
         parseRepositoryNwo(util.getRequiredEnvParam("GITHUB_REPOSITORY")),
         uploadResult.sarifID,
         getActionsLogger()
@@ -308,10 +307,7 @@ async function run() {
         `expect-error input was set to true but no error was thrown.`
       );
     }
-    core.exportVariable(
-      CODEQL_ACTION_ANALYZE_DID_COMPLETE_SUCCESSFULLY,
-      "true"
-    );
+    core.exportVariable(EnvVar.ANALYZE_DID_COMPLETE_SUCCESSFULLY, "true");
   } catch (unwrappedError) {
     const error = wrapError(unwrappedError);
     if (